# app/api/delete-message/route.ts
import { NextRequest, NextResponse } from "next/server"
import { createClient } from "@supabase/supabase-js"

const supabaseUrl = process.env.NEXT_PUBLIC_SUPABASE_URL || ""
const supabaseKey = process.env.SUPABASE_SERVICE_ROLE_KEY || ""
const supabase = createClient(supabaseUrl, supabaseKey)

export async function DELETE(req: NextRequest) {
  try {
    const { searchParams } = new URL(req.url)
    const messageId = searchParams.get("id")
    const userId = searchParams.get("user_id")

    if (!messageId || !userId) {
      return NextResponse.json({ error: "Faltan parámetros obligatorios" }, { status: 400 })
    }

    const { error } = await supabase
      .from("messages")
      .delete()
      .match({ id: messageId, user_id: userId })

    if (error) {
      return NextResponse.json({ error: error.message }, { status: 500 })
    }

    return NextResponse.json({ message: "Mensaje eliminado exitosamente" })
  } catch (error) {
    return NextResponse.json({ error: error instanceof Error ? error.message : "Error desconocido" }, { status: 500 })
  }
}