# swagger.yaml
openapi: 3.0.3
info:
  title: EduPlanner API
  description: API para la plataforma educativa EduPlanner - Planeaciones didácticas con IA
  version: 1.0.0
  contact:
    name: EduPlanner Support
    email: support@eduplanner.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000/api
    description: Servidor de desarrollo
  - url: https://eduplanner.vercel.app/api
    description: Servidor de producción
paths:
...
  /generate-pptx:
    post:
      tags:
        - Presentaciones
      summary: Generar archivo PowerPoint
      description: Genera un archivo PPTX basado en el contenido proporcionado
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                content:
                  type: string
                  description: Contenido para la presentación
      responses:
        '200':
          description: Archivo PPTX generado exitosamente
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '500':
          description: Error del servidor
...
  /delete-message:
    delete:
      tags:
        - Mensajes
      summary: Eliminar mensaje
      description: Elimina un mensaje específico de la base de datos
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                message_id:
                  type: string
              required:
                - message_id
      responses:
        '200':
          description: Mensaje eliminado exitosamente
        '400':
          description: ID del mensaje es requerido
        '500':
          description: Error del servidor
  /feedback:
    post:
      tags:
        - Feedback
      summary: Enviar retroalimentación
      description: Permite a los usuarios enviar comentarios y sugerencias
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                user_id:
                  type: string
                feedback:
                  type: string
                rating:
                  type: integer
                  minimum: 1
                  maximum: 5
              required:
                - feedback
      responses:
        '200':
          description: Feedback enviado exitosamente
        '400':
          description: Datos inválidos
        '500':
          description: Error del servidor
...
  /adapt-whatsapp:
    post:
      tags:
        - Integración
      summary: Adaptar contenido para WhatsApp
      description: Adapta el contenido educativo para ser enviado por WhatsApp
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                content:
                  type: string
                format:
                  type: string
                  enum: [text, image, document]
      responses:
        '200':
          description: Contenido adaptado exitosamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  adapted_content:
                    type: string
                  format:
                    type: string
        '500':
          description: Error del servidor
...
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token de autenticación JWT de Supabase
security:
  - BearerAuth: []
...